Do not set PROFILE_TASK environment variable
When this is set, the `--enable-optimizations` option is essentailly disabled,
as the profile task step is skipped.

For 3.13 on bullseye and alpine distros the version of expat has an issue that
causes the xml tests to fail and thus results in a failed build. This was not an
issue in version prior to 3.13 as the `Makefile` would ignore the failure and
continue as the default `PROFILE_TASK` included a `|| true` clause. That clause
was removed from 3.13's `Makefile`.

Thus for these distros, a patch is applied to the unit test to skip them